Next step forward. Config with properties copy and all the Stuff it needs to put the program in working state
The configuration staff is quite neat now. If the user types CONFIG he can maintain his setup. This means, giving it a source directory for the installed RoboCode version. Give it the number of wanted installations. Copy all files from source to destination (no need for an extra script anymore) and give it a robots source directory (can me more than one). As far i have tested it, is it fail resistant to a quite high level. When ever the user types something suspicious he gets a warning and can try it again.