Next step forward - challenges are now available
Challenges are now fully supported. The challenge can be changed and/or run again. With CHAL can the challenge be configured. The results are coming back from the processes and will be shown as output but the last step is still missing (next step).
Missing bots are copied to the runner directories and the challenger is saved  as well. Development bots are also available but this needs some further tweaks.