Next step forward - aspect driven scores
Well it sounds more than it is now but i still like it of some sort. The scorer can be written by the user himself and if he use the aspects everything will be calculated to his needs. This lets the opportunity open to write almost every kind of challenge result which comes from the overall battle results. To make statistics available it needs a little more effort.

The Aspects are not fully correct right now it is only the roundabout that counts. Next step will be to fill the aspects with the right values for the output