More result tweaking
The results can no be shown with all available scorer. If the user types 'res ?' he gets a selection of all available scorer and can choose what he wants to see. The scorer can be written while the program is running and it detects the new scorer on the fly. No need to restart the program.